fix: do not allow non C-contiguous buffer
They were not allowed in the C-extension but given a bug exists in PyPy, they could be accepted and return the wrong result.

Add a workaround for the PyPy bug and disallow non C-contiguous buffer in the pure python fallback implementation.

The `encodebytes` buffer check is also a bit stricter in the C-extension to mimic CPython behavior.